refactor: move out transcoding classes
