+++ src/Sinch/Conversation/Transcoding/TranscodeResponse.cs
@@ -1,0 +1,9 @@
+﻿using System.Collections.Generic;
+
+namespace Sinch.Conversation.Transcoding
+{
+    public class TranscodeResponse
+    {
+        public Dictionary<ConversationChannel, string> TranscodedMessage { get; set; }
+    }
+}
