+++ src/Sinch/Sinch.csproj
@@ -45,5 +45,8 @@
             <_Parameter1>DynamicProxyGenAssembly2</_Parameter1>
         </AssemblyAttribute>
     </ItemGroup>
+    <ItemGroup>
+      <Folder Include="Conversation\Capability\" />
+    </ItemGroup>
 
 </Project>
